--- src/main/java/com/example/openapi/repository/SearchResult.java
@@ -1,14 +1,0 @@
-package com.example.openapi.repository;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SearchResult {
-    private String title;
-    private String link;
-    private String description;
-    private String bloggername;
-    private String bloggerlink;
-}
