+++ src/main/resources/application.properties
@@ -1,6 +1,7 @@
-naver.openapi.search-url=https://openapi.naver.com/v1/search/blog.json
-naver.openapi.clientId=-
-naver.openapi.clientSecret=-
+naver.openapi.blog-url=https://openapi.naver.com/v1/search/blog.json
+naver.openapi.movie-url=https://openapi.naver.com/v1/search/movie.json
+naver.openapi.clientId=
+naver.openapi.clientSecret=
 spring.profiles.active=local
 
 server.port=8081
